Move the filename regex ignore test to the glob integration test.
